Added in function to load data from folder
Added in a preload image folder for scooter data. The data is structured as measurement files for left,right and center and images from all 3 cameras in a folder. The function matches the individual frames with the individual measurement files. This allows the frames to not be synced with each other. 

This is different compared to the CIL data format whereby all 3 frames (left,mid,right) are synced to one measurement document.